+++ packages/parser/CHANGELOG.md
@@ -1,5 +1,7 @@
 # CHANGELOG
 
+- Removed React 15 support.
+
 ### 1.0.0
 
 - Initial public release.