+++ packages/provider/CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+- Fixed a race condition where fetching would still trigger setState when the
+  component was unmounted already.
+- Removed React 15 support.
+
 ### 1.1.1
 
 - Allowed the `url` function that was added in 1.1.0 to access the provider